Reorganize compare components
