+++ src/app/manager-dashboard/compare/compare-courses.component.html
