+++ src/app/manager-dashboard/compare/compare-resources.component.html
