+++ src/app/manager-dashboard/compare/compare.component.ts
