+++ src/app/manager-dashboard/manger-dashboard-router.module.ts
@@ -7,7 +7,7 @@ import { ManagerDashboardConfigurationComponent } from './manager-dashboard-conf
 import { ReportsComponent } from './reports/reports.component';
 import { ReportsDetailComponent } from './reports/reports-detail.component';
 import { ReportsPendingComponent } from './reports/reports-pending.component';
-import { CompareComponent } from './compare.component';
+import { CompareComponent } from './compare/compare.component';
 
 const routes: Routes = [
   { path: '', component: ManagerDashboardComponent },
